Added searching and host settings
When you search for sessions it now generates a list of the sessions available to join. When you host a session, you can now feed in the number of players the lobby should be able to hold and whether this server is LAN or not. There is also the character selection code which is currently being made.